+++ lib/rubocop/cop/style/trivial_accessors.rb
@@ -32,6 +32,10 @@ module Rubocop
           cop_config['AllowPredicates']
         end
 
+        def allow_dsl_writers?
+          cop_config['AllowDSLWriters']
+        end
+
         def whitelist
           whitelist = cop_config['Whitelist']
           Array(whitelist).map(&:to_sym) + [:initialize]
@@ -41,9 +45,14 @@ module Rubocop
           method_name[-1] == '?'
         end
 
+        def dsl_writer?(method_name)
+          !method_name.to_s.end_with?('=')
+        end
+
         def trivial_reader?(method_name, args, body)
           looks_like_trivial_reader?(args, body) &&
-            !allowed_method?(method_name, body)
+            !allowed_method?(method_name, body) &&
+            !allowed_reader?(method_name)
         end
 
         def looks_like_trivial_reader?(args, body)
@@ -51,20 +60,28 @@ module Rubocop
         end
 
         def trivial_writer?(method_name, args, body)
-          looks_like_trivial_writer?(args, body) &&
-            !allowed_method?(method_name, body)
+          looks_like_trivial_writer?(method_name, args, body) &&
+            !allowed_method?(method_name, body) &&
+            !allowed_writer?(method_name)
         end
 
-        def looks_like_trivial_writer?(args, body)
+        def looks_like_trivial_writer?(method_name, args, body)
           args.children.size == 1 && args.children[0].type != :restarg &&
             body && body.type == :ivasgn &&
             body.children[1] && body.children[1].type == :lvar
         end
 
         def allowed_method?(method_name, body)
-          allow_predicates? && predicate?(method_name) ||
-            whitelist.include?(method_name) ||
+          whitelist.include?(method_name) ||
             exact_name_match? && !names_match?(method_name, body)
+        end
+
+        def allowed_writer?(method_name)
+          allow_dsl_writers? && dsl_writer?(method_name)
+        end
+
+        def allowed_reader?(method_name)
+          allow_predicates? && predicate?(method_name)
         end
 
         def names_match?(method_name, body)